requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.3            # via django
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
coverage==5.0.3           # via surveys_api (setup.py)
django==3.0.3             # via djangorestframework, drf-yasg, surveys_api (setup.py)
djangorestframework==3.11.0  # via drf-extensions, drf-yasg, surveys_api (setup.py)
drf-extensions==0.6.0     # via surveys_api (setup.py)
drf-yasg==1.17.1          # via surveys_api (setup.py)
factory-boy==2.12.0       # via surveys_api (setup.py)
faker==4.0.1              # via factory-boy, surveys_api (setup.py)
idna==2.9                 # via requests
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
markupsafe==1.1.1         # via jinja2
packaging==20.1           # via drf-yasg
psycopg2==2.8.4           # via surveys_api (setup.py)
pyparsing==2.4.6          # via packaging
python-dateutil==2.8.1    # via faker
python-decouple==3.3      # via surveys_api (setup.py)
pytz==2019.3              # via django
requests==2.23.0          # via coreapi
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
six==1.14.0               # via drf-yasg, packaging, python-dateutil
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.8           # via requests
